Added utilites to make and animate tilemaps.

+++ genutils.py
@@ -20,8 +20,8 @@ def cart_to_polar(xv,yv,maxrad=1,minrad=0,axis=0,phase=0):
     return((xnew_v,ynew_v))
 
 def sample_line(pos1,pos2,nsample):
-    inc_x=int((pos2[0]-pos1[0])/(nsample-1))
-    inc_y=int((pos2[1]-pos1[1])/(nsample-1))
+    inc_x=(pos2[0]-pos1[0])/(nsample-1)
+    inc_y=(pos2[1]-pos1[1])/(nsample-1)
     pos_all=[pos1]
     for i in range(nsample-2):
         pos_all.append((int(pos1[0]+(i+1)*inc_x),int(pos1[1]+(i+1)*inc_y)))
